Refactored `magefiles/magefile.go` for sustainable upkeep.
**Added:**

- `docgen.go` - Extracted documentation generation logic from `magefile.go` to `docgen.go`.
- `testing.go` - Segregated testing-related functions from `magefile.go` to `testing.go`.

**Changed:**

- `magefile.go` - Split existing code into multiple files to promote easier maintenance and better organization.
